API: KmerFrequencySpectra

Types

KmerAnalysis.KmerFrequencySpectra — Type

A type for storing a frequency histogram of MerCounts, also sometimes referred to as a Kmer spectra.

KmerAnalysis.SpectraPeak — Type

A simple struct describing the location and significance (or persistence) of a peak detected in kmer frequency spectra data.

It has the following fields:

peak: The k-mer frequency at which the peak is at its highest.

left: The k-mer frequency constituting the left bound of the peak.

right: The k-mer frequency constituting the right bound of the peak.

persistence: The persistence of the peak in persistent homology analysis - essentially a significance value for the peak. Higher value = more significant.

Public / Safe methods

KmerAnalysis.spectra — Function

spectra(freqs::Vector{MerCount{M}}, min_count::Integer) where {M<:AbstractMer}

Build a 1 dimensional kmer frequency spectra, from a vector of kmer counts, excluding any kmer counts that don't meet min_count.

spectra(xfreqs::Vector{MerCount{M}}, yfreqs::Vector{MerCount{M}}, min_count::Integer = 0) where {M<:AbstractMer}

Build

spectra(x::Pair{I,C}, min_count::Integer = 0) where {I,C<:AbstractKmerCounter}

Build a 1 dimensional kmer frequency spectra, using a pair of (input data => kmer counter) as the input argument x. The spectra function uses the kmer counter to get the counts from the input data, before computing the spectra.

Any kmer counts that don't meet min_count are excluded.

KmerAnalysis.find_spectra_peaks — Function

find_spectra_peaks(s::KmerFrequencySpectra{1})

Automatically detect all the peaks present in the signal of a 1D k-mer frequency spectra, using a persistent homology method.

Returns a vector of SpectraPeak.

Internal / Unsafe methods

Missing docstring.

Missing docstring for build_persistent_homology. Check Documenter's build log for details.