Counting kmers in read datasets

MerCounting provides some more dedicated counting algorithms for counting kmers in read datasets, particularly ReadDatastores.

This is because as conceptually simple as counting kmers is, to do it quickly for large sequencing read datasets output by sequencing machines can be difficult for some datasets.

There are many ways you can try to optimise the kmer counting process, and many kmer counting tools already exist.

MerCounting provides a Counters submodule, which contains nessecery methods and types required to implement various kinds of kmer counter, as well as exporting a selection of "off-the-shelf" methods that use different counting strategies, one or several of which hopefully suit the user's dataset and computational resources available. We describe and showcase these below.

The serial_mem counter

The serial_mem counter is the simplest kmer counter for reads that is provided.

It counts kmers in each read serially, and stores the kmers and counts entirely in RAM.

MerCounting.Counters.serial_mem — Function

serial_mem(::Type{M}, input::DatastoreBuffer{<:ReadDatastore}, count_mode::CountMode, range::AbstractRange = 1:length(input)) where {M<:AbstractMer}

MerCounting's simplest kmer counting method.

Build a sorted list (vector) of kmer counts (MerCount), serially and entirely in memory.

Warning

This function is a serial and in memory MerCount list builder that can build a kmer count from a ReadsDatastore on its own (if you have memory and time), but it is also intended to be composed into other multi-process or multi-threaded kmer counting strategies.

This method estimates roughly how many kmers will be generated by the reads specified by range in the dataset. It then pre-allocates an array to contain them. It then collects the kmers, sorts, them, and then collapses them into a list of counts sorted by the kmer.

So if you want to count kmers and have the resources to throw at it, this is the simplest method, and possibly even the quickest given that simplicity.

serial_mem(::Type{M}, input::ReadDatastore, count_mode::CountMode, range::AbstractRange = 1:length(input)) where {M<:AbstractMer}

MerCounting's simplest kmer counting method.

Build a sorted list (vector) of kmer counts (MerCount), serially and entirely in memory.

Warning

This function is a serial and in memory MerCount list builder that can build a kmer count from a ReadsDatastore on its own (if you have memory and time), but it is also intended to be composed into other multi-process or multi-threaded kmer counting strategies.

This method estimates roughly how many kmers will be generated by the reads specified by range in the dataset. It then pre-allocates an array to contain them. It then collects the kmers, sorts, them, and then collapses them into a list of counts sorted by the kmer.

So if you want to count kmers and have the resources to throw at it, this is the simplest method, and possibly even the quickest given that simplicity.

It is fairly simple to use. First you need a ReadDatastore, if you need to recall how to create one then head over to here in ReadDatastores.jl's documentation.

The example below opens a datastore before using serial_mem to count the kmers in the read datastore.

julia> using MerCounting

julia> using ReadDatastores

julia> ds = @openreads "ecoli-test-paired.prseq"
ERROR: LoadError: SystemError: opening file "ecoli-test-paired.prseq": No such file or directory
in expression starting at none:1

julia> kl = Counters.serial_mem(DNAMer{31}, ds, CANONICAL)
ERROR: UndefVarError: ds not defined