Counting kmers in read datasets

MerCounting provides some more dedicated counting algorithms for counting kmers in read datasets, particularly ReadDatastores.

This is because as conceptually simple as counting kmers is, to do it quickly for large sequencing read datasets output by sequencing machines can be difficult for some datasets.

There are many ways you can try to optimise the kmer counting process, and many kmer counting tools already exist.

MerCounting provides a Counters submodule, which contains nessecery methods and types required to implement various kinds of kmer counter, as well as exporting a selection of "off-the-shelf" methods that use different counting strategies, one or several of which hopefully suit the user's dataset and computational resources available. We describe and showcase these below.

The serial_mem counter

The serial_mem counter is the simplest kmer counter for reads that is provided.

It counts kmers in each read serially, and stores the kmers and counts entirely in RAM.

KmerAnalysis.Counters.serial_mem — Function

serial_mem(::Type{M}, input::DatastoreBuffer{<:ReadDatastore}, count_mode::CountMode, range::AbstractRange = 1:length(input)) where {M<:AbstractMer}

MerCounting's simplest kmer counting method.

Build a sorted list (vector) of kmer counts (MerCount), serially and entirely in memory.

Warning

This function is a serial and in memory MerCount list builder that can build a kmer count from a ReadsDatastore on its own (if you have memory and time), but it is also intended to be composed into other multi-process or multi-threaded kmer counting strategies.

This method estimates roughly how many kmers will be generated by the reads specified by range in the dataset. It then pre-allocates an array to contain them. It then collects the kmers, sorts, them, and then collapses them into a list of counts sorted by the kmer.

So if you want to count kmers and have the resources to throw at it, this is the simplest method, and possibly even the quickest given that simplicity.

serial_mem(::Type{M}, input::ReadDatastore, count_mode::CountMode, range::AbstractRange = 1:length(input)) where {M<:AbstractMer}

MerCounting's simplest kmer counting method.

Build a sorted list (vector) of kmer counts (MerCount), serially and entirely in memory.

Warning

This function is a serial and in memory MerCount list builder that can build a kmer count from a ReadsDatastore on its own (if you have memory and time), but it is also intended to be composed into other multi-process or multi-threaded kmer counting strategies.

This method estimates roughly how many kmers will be generated by the reads specified by range in the dataset. It then pre-allocates an array to contain them. It then collects the kmers, sorts, them, and then collapses them into a list of counts sorted by the kmer.

So if you want to count kmers and have the resources to throw at it, this is the simplest method, and possibly even the quickest given that simplicity.

It is fairly simple to use. First you need a ReadDatastore, if you need to recall how to create one then head over to here in ReadDatastores.jl's documentation.

The example below opens a datastore before using serial_mem to count the kmers in the read datastore.

julia> using KmerAnalysis, ReadDatastores

julia> ds = @openreads "ecoli-test-paired.prseq"
Paired Read Datastore 'my-ecoli-test': 20 reads (10 pairs)

julia> kl = Counters.serial_mem(DNAMer{31}, ds, CANONICAL)
4916-element Array{MerCount{BioSequences.Mer{BioSequences.DNAAlphabet{2},31}},1}:
 AAAAAAAAGTAATCGTCGGCATGTCCGGCGG occurs 1 times
 AAAAAAAGTAATCGTCGGCATGTCCGGCGGT occurs 1 times
 AAAAAAATAAAATTTAGTGCTGTACAGAGCG occurs 1 times
 AAAAAAGTAATCGTCGGCATGTCCGGCGGTG occurs 1 times
 AAAAAATAAAATTTAGTGCTGTACAGAGCGC occurs 1 times
 AAAAAATCCACCAGCAGCAGGTTGATGTCAG occurs 1 times
 AAAAAATCGACCAGGAGCGGTTGGATGACAG occurs 1 times
 AAAAAATCGTTGCGCTTTTTCTCGCCGCCGG occurs 1 times
 AAAAACATCGCGGGTGTTGAAGAAAAATTAA occurs 1 times
 AAAAACCAACGTTTACCGAACGGGTTAATAA occurs 2 times
 ⋮
 TTCGGCGTGCGACCGGCTTTATATTCGGCAA occurs 1 times
 TTCTGGCTGAAGAGGTGGTGGAAATTCCCAA occurs 1 times
 TTCTTGCTTAAGAGGTGGTGGAAATTCCCAA occurs 1 times
 TTGCCATTACCTATGTCTACAAAGGTGACAA occurs 1 times
 TTGCGAACAATTATTCCCCATGCTTCAGCAA occurs 1 times
 TTGCGCTTTTTCTCGCCGCCGGAAAAACCAA occurs 2 times
 TTGGCAGCATCTTCTTTGGTGGTTGCACCAA occurs 1 times
 TTGTCCCAACACATGTGCCAGCCGATAAAAA occurs 1 times
 TTTGATTTTCAGGATTTGATGGAAGAGAAAA occurs 1 times

The dist_mem counter

The dist_mem counter is a multi-process version of serial_mem.

KmerAnalysis.Counters.dist_mem — Function

The multi-process parallel version of the serial_mem kmer counter.

Like serial_mem this counter does all counting and processing in memory, so it is not designed to limit memory use, but if using a cluster of machines the memory used per worker machine would be less. Of course the data is copied back over to the master process and the results combined and so the master node has to have enough RAM available to hold all the result.

It is used in a very similar fashion to serial_mem, but you need to have multiple julia worker processes available to gain any benefit.

julia> # Spin up some extra julia worker processes, make sure they are using the
       # appropriate Project.toml file for your project.
       using Distributed

julia> addprocs(8, exeflags="--project=../../../docs/")
8-element Array{Int64,1}:
 2
 3
 4
 5
 6
 7
 8
 9

julia> # Make sure the packages are loaded on all the workers.
       @everywhere using KmerAnalysis, ReadDatastores, BioSequences

julia> ds = @openreads "ecoli-test-paired.prseq"
Paired Read Datastore 'my-ecoli-test': 20 reads (10 pairs)

julia> kl = Counters.dist_mem(DNAMer{31}, ds, CANONICAL)
[ Info: Splitting all reads in ecoli-test-paired.prseq, across 8 worker processes and counting kmers
[ Info: Waiting for workers to finish
[ Info: Merging all counts from 8 worker processes
4916-element Array{MerCount{Mer{DNAAlphabet{2},31}},1}:
 AAAAAAAAGTAATCGTCGGCATGTCCGGCGG occurs 1 times
 AAAAAAAGTAATCGTCGGCATGTCCGGCGGT occurs 1 times
 AAAAAAATAAAATTTAGTGCTGTACAGAGCG occurs 1 times
 AAAAAAGTAATCGTCGGCATGTCCGGCGGTG occurs 1 times
 AAAAAATAAAATTTAGTGCTGTACAGAGCGC occurs 1 times
 AAAAAATCCACCAGCAGCAGGTTGATGTCAG occurs 1 times
 AAAAAATCGACCAGGAGCGGTTGGATGACAG occurs 1 times
 AAAAAATCGTTGCGCTTTTTCTCGCCGCCGG occurs 1 times
 AAAAACATCGCGGGTGTTGAAGAAAAATTAA occurs 1 times
 AAAAACCAACGTTTACCGAACGGGTTAATAA occurs 2 times
 ⋮
 TTCGGCGTGCGACCGGCTTTATATTCGGCAA occurs 1 times
 TTCTGGCTGAAGAGGTGGTGGAAATTCCCAA occurs 1 times
 TTCTTGCTTAAGAGGTGGTGGAAATTCCCAA occurs 1 times
 TTGCCATTACCTATGTCTACAAAGGTGACAA occurs 1 times
 TTGCGAACAATTATTCCCCATGCTTCAGCAA occurs 1 times
 TTGCGCTTTTTCTCGCCGCCGGAAAAACCAA occurs 2 times
 TTGGCAGCATCTTCTTTGGTGGTTGCACCAA occurs 1 times
 TTGTCCCAACACATGTGCCAGCCGATAAAAA occurs 1 times
 TTTGATTTTCAGGATTTGATGGAAGAGAAAA occurs 1 times

Naturally, for such a small and simple example as this, using dist_mem for such a small dataset is probably just worse than doing it in memory serially, because of the overheads. For real datasets you should see a benefit.